+++ exercicios/q04.c
@@ -3,11 +3,11 @@
 // função que converte real para dolar 
 int main()
 {
-    double valor, cotacao;
+    float valor, cotacao;
     printf("Digite o valor que voce quer converter de real para dolar:\n");
-    scanf("%lf", &valor);
+    scanf("%f", &valor);
     printf("Digite a cotacao atual:\n");
-    scanf("%lf", &cotacao);
-    printf("%.2lf$ eh o cambio de real para dolar!", valor / cotacao);
+    scanf("%f", &cotacao);
+    printf("%.2f$ eh o cambio de real para dolar!", valor / cotacao);
     return 0;
 }
